# apps/disher-backend-3.0/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Disher API",
    "version": "1.0.0",
    "description": "Backend for the Disher food diary. Errors share one RFC 9457-subset problem+json shape — see api/errors.ts."
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-disher1.session_token"
      }
    },
...
  },
  "paths": {
...
    "/api/user-reports/": {
      "post": {
        "operationId": "submitUserReport",
        "tags": [
          "feedback"
        ],
        "description": "User-facing «Сообщить о проблеме» sink — text + client metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "title": "UserReport",
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "page": {
                    "type": "string"
                  },
                  "screenSize": {
                    "type": "string"
                  },
                  "userAgent": {
                    "type": "string"
                  },
                  "pwa": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/balance/ledger": {
      "get": {
        "operationId": "listLedger",
        "tags": [
          "billing"
        ],
        "description": "Wallet ledger entries, newest first.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/admin/me": {
      "get": {
        "operationId": "getAdminMe",
        "tags": [
          "admin"
        ],
        "description": "Probe: 200 means the caller is an admin. Env-bootstrap admins have role='user', so the client cannot tell from the session alone.",
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "listUsers",
        "tags": [
          "admin"
        ],
        "description": "Every user + wallet balance. No pagination — the client filters locally.",
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/users/{id}/topup": {
      "post": {
        "operationId": "topupUser",
        "tags": [
          "admin"
        ],
        "description": "Credit a user's wallet. Idempotent per requestId — a repeat returns the first result with alreadyApplied:true.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "title": "TopupRequest",
                "type": "object",
                "properties": {
                  "amountKop": {
                    "description": "Positive integer, kopecks."
                  },
                  "reason": {
                    "description": "Non-empty, <= 200 chars."
                  },
                  "requestId": {
                    "description": "Idempotency key, 1..100 chars. 'welcome' is reserved. A repeat returns alreadyApplied:true."
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/auth-events": {
      "get": {
        "operationId": "listAuthEvents",
        "tags": [
          "admin"
        ],
        "description": "Login diagnostics, newest first. `problems=1` narrows to non-success; `q` matches an email substring or a user id.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Clamped to 1..500 by the handler; default 100."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "problems",
            "required": false,
            "description": "'1' = failures only."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/user-reports": {
      "get": {
        "operationId": "listUserReports",
        "tags": [
          "admin"
        ],
        "description": "Production bug reports (table user_reports) joined with the author's email, newest first. `q` matches email / text / user id.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Clamped to 1..500 by the handler; default 100."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/users/{id}/ledger": {
      "get": {
        "operationId": "getUserLedger",
        "tags": [
          "admin"
        ],
        "description": "One user's ledger, newest first, including the meta that carries the topup reason.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}